Add minimum controllers
This commit adds the minimum amount of controllers required by this
project, which are AirConditionerController, AlarmController, and
BlindsController. AirConditionerController controls the AC based on the
temperature of the surroundings and time. AlarmController will currently
sound an alarm if the wind speed is too high. Likewise, BlindsController
will close the blinds if the wind is too high and/or it is already
evening.