Change AirConditionerController
This commit aims such that on the main function, the sensor and the
controller will be separate objects. For example, with AC Controller, first we set the parameters
of the thermometer, then as the temperature have been acquired, we can
use the value obtained for the controller to do appropriate actions.