Revision to the GUI.
The problem with the GUI previously created is because the
frame.setVisible() is set BEFORE the components are added to the panel,
and before the panel is added to the frame. The frame.setVisible should
be added after all the components are added to the frame. This commit
also finally make the temperature slider functional.