Automate the changes in wind speed
This commit simulates what the project wants to be, but in GUI form.
Basically, there will be a random number from -3 to 3 to change the
current windSpeed. Then, the windSpeed obtained from the sensor will be
passed into the controller, which then will control the actor. To
simulate 1 second, there will be a thread which will sleep for 1000
milliseconds everytime the function ends.